Updated DCF parsing to handle multiple ID items per table
DHS HIV survey datasets define three ID columns (as opposed to a single
ID column CASEID which may or may not include HHID and individual ID
within a single string). The code needed to be updated as otherwise only
the last of these ID columns would be written to the specification file
and thus extracted from the data file.