Made range expansion strategy a parameter, fixed bug when [NA] is in label
The chunk-matching code is based on looking for certain keywords
enclosed in [square] brackets. Any text enclosed in square brackets but
not matching a keyword led to that chunk being skipped.
Unfortunately some files contain [NA] in the label text and this led to
those items not being recorded. However the values were still recorded
and were added to the previous item that was actually recorded! The
string similarity checking may have flagged this sometimes, but it only
issues a warning.