Updated DCF parser to handle non-integer value ranges
Newer DCF files (Phase 7 surveys) seem to contain value ranges whose top
and/or bottom values are non-integer. The parsing code uses a regular
expression to pull out value ranges limits and this did not handle
decimal numbers. Now fixed. If a range top or bottom value is not a
whole number, the range will not be expanded.